moved autoreply commands to messages module
